Correct processing of slow-scroll.
We now correctly delete the additional line that the slow scrolling leaves
at the top of the scroll region while the region shifted by a few pixels.

The hardware deletes this line by removing `scroll_latch` pixels at the
top of the region and resetting the scroll to the start of the next text
line at the bottom. We do it in one go at the top.